chore: restatructure frontend js
