+++ frontend/index.html
@@ -152,6 +152,7 @@
   <!-- Translation Files -->
   <script src="translations/en-US.js"></script>
   <script src="translations/th-TH.js"></script>
-  <script src="js/script.js"></script>
+  <!-- Main modular JavaScript -->
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
