+++ frontend/js/notification.js
@@ -1,0 +1,33 @@
+// Notification system
+export const notificationManager = {
+  show(message, type = 'info') {
+    const notification = document.createElement('div');
+    notification.className = `notification notification-${type}`;
+    notification.innerHTML = `
+      <span>${message}</span>
+      <button class="notification-close">&times;</button>
+    `;
+
+    document.body.appendChild(notification);
+
+    // Show notification
+    setTimeout(() => notification.classList.add('show'), 100);
+
+    // Auto remove after 5 seconds
+    setTimeout(() => this.remove(notification), 5000);
+
+    // Manual close
+    notification.querySelector('.notification-close').addEventListener('click', () => {
+      this.remove(notification);
+    });
+  },
+
+  remove(notification) {
+    notification.classList.remove('show');
+    setTimeout(() => {
+      if (notification.parentNode) {
+        notification.parentNode.removeChild(notification);
+      }
+    }, 300);
+  }
+};
